+++ audit-triplestore/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -10,11 +10,12 @@
    <!-- OSGI blueprint property placeholder -->
    <cm:property-placeholder persistent-id="org.fcrepo.camel.audit" update-strategy="reload">
      <cm:default-properties>
+       <cm:property name="error.maxRedeliveries" value="10"/>
        <cm:property name="jms.brokerUrl" value="tcp://localhost:61616"/>
        <cm:property name="jms.fcrepoEndpoint" value="topic:fedora"/>
-       <cm:property name="triplestore.baseUrl" value="localhost:3030/test/update?"/>
+       <cm:property name="triplestore.baseUrl" value="localhost:3030/test/update"/>
        <!-- Base URI to be used in contructing the URI for the JMS event -->
-       <cm:property name="event.baseUri" value="http://example.com"/>
+       <cm:property name="event.baseUri" value="http://example.com/event"/>
      </cm:default-properties>
    </cm:property-placeholder>
 
@@ -24,7 +25,6 @@
   </bean>
 
   <camelContext xmlns="http://camel.apache.org/schema/blueprint">
-
     <package>org.fcrepo.camel.audit.triplestore</package>
   </camelContext>
